android_hardware_u8500

+++ media/libomxil-bellagio/src/base/omx_base_sink.c
@@ -24,7 +24,8 @@
   02110-1301  USA
 
 */
-
+#include <sys/syscall.h>
+#include <unistd.h>
 #include <omxcore.h>
 #include <omx_base_sink.h>
 
@@ -316,7 +317,7 @@ void* omx_base_sink_twoport_BufferMgmtFunction (void* param) {
       }
     }
 
-    for(i=0;i < (omx_base_component_Private->sPortTypesParam[OMX_PortDomainAudio].nPorts  +
+    for(i=0;(unsigned int)i < (omx_base_component_Private->sPortTypesParam[OMX_PortDomainAudio].nPorts  +
                  omx_base_component_Private->sPortTypesParam[OMX_PortDomainVideo].nPorts +
                  omx_base_component_Private->sPortTypesParam[OMX_PortDomainImage].nPorts +
                  omx_base_component_Private->sPortTypesParam[OMX_PortDomainOther].nPorts);i++) {
